Add password matching check
Added a password matching check using JavaScript and CSS. If the
password fields are empty or mismatched, a red border is given to the
password fields and a red error message is displayed. If the passwords
match, an error class is removed to remove the red border and error
message.